# services/mobility/src/main/java/com/city/mobility/dto/TransportDTO.java
package com.city.mobility.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class TransportDTO {
    private Long id;
    private String lineNumber;
    private String type;
    private String origin;
    private String destination;
    private LocalDateTime departureTime;
    private LocalDateTime arrivalTime;
    private String status;
    private Integer delayMinutes;
    private Integer availableSeats;
}